% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstorr.R
\name{fstorr}
\alias{fstorr}
\title{Create an fstorr object}
\usage{
fstorr(root, get_resource, hash_content = NULL)
}
\arguments{
\item{root}{Root to store files in - a path to a directory.  If it
exists it must be a directory and you should treat the contents
as opaque.}

\item{get_resource}{A function that gets a resource.  It must take
as its argument a key, and it must return a raw vector
(character vectors are not supported on return because they are
complicated for thinking about line-endings and hashes).}

\item{hash_content}{A function that hashes a resource.  If not
given then \code{openssl::sha256} will be used.  If providing
your own it must take a filename as an argument and return a
character string corresponding to a hash.  It is \emph{very
important} that the same mapping function here is used for all
creations of an \code{fstorr} object!}
}
\value{
A \code{fstorr} object
}
\description{
Create an \code{fstorr} object - a place to store data by content
based on a global key.
}
\examples{
path <- tempfile()
obj <- fstorr::fstorr(path, charToRaw)
obj$list()
obj$get("a")
}
\author{
Rich Fitzjohn
}
